backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.8.3; python_version >= '3.6'
aiosignal==1.2.0; python_version >= '3.6'
asgiref==3.5.2; python_version >= '3.7'
astroid==2.12.12; python_full_version >= '3.7.2'
asttokens==2.0.5
async-timeout==4.0.2; python_version >= '3.6'
attrs==22.1.0; python_version >= '3.5'
base58==2.1.1; python_version >= '3.5'
bitarray==2.6.0
black==22.10.0
certifi==2022.9.24; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
cytoolz==0.12.0; python_version >= '3.5'
dataclassy==0.11.1; python_version >= '3.6'
dill==0.3.6; python_version >= '3.7'
django-cors-headers==3.13.0
django-filter==22.1
django==4.1.3
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
eip712==0.1.0; python_version >= '3.6' and python_version < '4'
eth-abi==2.2.0; python_version >= '3.6' and python_version < '4'
eth-account==0.5.9; python_version >= '3.6' and python_version < '4'
eth-brownie==1.19.2
eth-event==1.2.3; python_version >= '3.6' and python_version < '4'
eth-hash[pycryptodome]==0.3.3; python_version >= '3.5' and python_version < '4'
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1; python_version >= '3.6' and python_version < '4'
eth-typing==2.3.0; python_version >= '3.5' and python_version < '4'
eth-utils==1.10.0; python_version >= '3.5' and python_version < '4' and python_full_version != '3.5.2'
execnet==1.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
frozenlist==1.3.1; python_version >= '3.7'
hexbytes==0.2.3; python_version >= '3.6' and python_version < '4'
hypothesis==6.27.3; python_version >= '3.6'
idna==3.4; python_version >= '3.5'
importlib-metadata==5.0.0; python_version < '3.10'
inflection==0.5.0; python_version >= '3.5'
iniconfig==1.1.1
ipfshttpclient==0.8.0a2; python_full_version >= '3.6.2' and python_full_version not in '3.7.0, 3.7.1'
isort==5.10.1; python_version < '4' and python_full_version >= '3.6.1'
jsonschema==3.2.0
lazy-object-proxy==1.7.1; python_version >= '3.6'
lru-dict==1.1.8
markdown==3.4.1
mccabe==0.7.0; python_version >= '3.6'
multiaddr==0.0.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
multidict==6.0.2; python_version >= '3.7'
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
packaging==21.3; python_version >= '3.6'
parsimonious==0.8.1
pathspec==0.10.1; python_version >= '3.7'
platformdirs==2.5.2; python_version >= '3.7'
pluggy==1.0.0; python_version >= '3.6'
prompt-toolkit==3.0.31; python_full_version >= '3.6.2'
protobuf==3.19.5; python_version >= '3.5'
psutil==5.9.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
psycopg2-binary==2.9.5
py-solc-ast==1.2.9; python_version >= '3.6' and python_version < '4'
py-solc-x==1.1.1; python_version >= '3.6' and python_version < '4'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycryptodome==3.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pygments-lexer-solidity==0.7.0
pygments==2.13.0; python_version >= '3.6'
pyjwt==1.7.1
pylint==2.15.5
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyrsistent==0.18.1; python_version >= '3.7'
pytest-forked==1.4.0; python_version >= '3.6'
pytest-xdist==1.34.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pytest==6.2.5; python_version >= '3.6'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.16.0
pythx==1.6.1
pytz==2022.6
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests==2.28.1; python_version >= '3.7' and python_version < '4'
rlp==2.0.1
semantic-version==2.10.0; python_version >= '2.7'
setuptools==65.5.1; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sortedcontainers==2.4.0
sqlparse==0.4.3; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version < '3.11'
tomlkit==0.11.6; python_version >= '3.6'
toolz==0.12.0; python_version >= '3.5'
tqdm==4.64.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.4.0; python_version < '3.10'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
varint==1.0.2
vvm==0.1.0; python_version >= '3.6' and python_version < '4'
vyper==0.3.7; python_version < '3.11' and python_version >= '3.7'
wcwidth==0.2.5
web3==5.31.1; python_version >= '3.6' and python_version < '4'
websockets==9.1; python_full_version >= '3.6.1'
wheel==0.37.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
yarl==1.8.1; python_version >= '3.7'
zipp==3.10.0; python_version >= '3.7'
